+++ src/Exception/InvalidRowException.php
@@ -4,13 +4,15 @@
 *
 * @license http://opensource.org/licenses/MIT
 * @link https://github.com/thephpleague/csv/
-* @version 7.2.0
+* @version 8.0.0
 * @package League.csv
 *
 * For the full copyright and license information, please view the LICENSE
 * file that was distributed with this source code.
 */
 namespace League\Csv\Exception;
+
+use InvalidArgumentException;
 
 /**
  *  Thrown when a data is not validated prior to insertion
@@ -19,16 +21,18 @@ namespace League\Csv\Exception;
  * @since  7.0.0
  *
  */
-class InvalidRowException extends \InvalidArgumentException
+class InvalidRowException extends InvalidArgumentException
 {
     /**
      * Validator which did not validated the data
+     *
      * @var string
      */
     private $name;
 
     /**
      * Validator Data which caused the error
+     *
      * @var array
      */
     private $data;